Move metadata to ClientModel itself
